requirements.txt
asgiref==3.4.1
autopep8==1.6.0
Django==3.2.11
django-cors-headers==3.10.1
django-rest-framework==0.1.0
djangorestframework==3.13.1
gunicorn==20.1.0
Pillow==9.0.0
pycodestyle==2.8.0
pytz==2021.3
sqlparse==0.4.2
toml==0.10.2
whitenoise==5.3.0
